remove Swap and AddTerms, reorganize files
